This Instagram test is failing locally
The URL requests are being redirected to a login page and thus the
expected payload does not exist on the page being received. It seems to
be a IP based redirect due to the number of requests. So it is
commented out until a better solution can be put into place.